made auth rate limit configfurable

+++ src/Http/Controllers/AuthController.php
@@ -81,8 +81,8 @@ class AuthController extends Controller
 
         // Per-account (email+IP) AND per-IP throttles. The IP cap blocks
         // user-enumeration / credential-spraying across many accounts from one source.
-        $this->ensureNotRateLimited($rlKey, config('jasmine.auth.rate_limits.login.attempts', 5));
-        $this->ensureNotRateLimited($ipKey, config('jasmine.auth.rate_limits.login.ip_attempts', 30));
+        $this->ensureNotRateLimited($rlKey, (int)config('jasmine.auth.rate_limits.login.attempts', 5));
+        $this->ensureNotRateLimited($ipKey, (int)config('jasmine.auth.rate_limits.login.ip_attempts', 30));
 
         if (static::guard()->attempt($credentials, $request->boolean('remember'))) {
             RateLimiter::clear($rlKey);
@@ -91,8 +91,8 @@ class AuthController extends Controller
             return Inertia::location(session('url.intended', route('jasmine.dashboard')));
         }
 
-        RateLimiter::hit($rlKey, config('jasmine.auth.rate_limits.login.decay', 60));
-        RateLimiter::hit($ipKey, config('jasmine.auth.rate_limits.login.ip_decay', 60));
+        RateLimiter::hit($rlKey, (int)config('jasmine.auth.rate_limits.login.decay', 60));
+        RateLimiter::hit($ipKey, (int)config('jasmine.auth.rate_limits.login.ip_decay', 60));
 
         return back()->withErrors(['email' => trans('auth.failed')])->onlyInput('email');
     }
@@ -222,8 +222,8 @@ class AuthController extends Controller
         $rlKey = $this->throttleKey('forgot', $data['email'], $ip);
         $ipKey = $this->ipThrottleKey('forgot', $ip);
 
-        $this->ensureNotRateLimited($rlKey, config('jasmine.auth.rate_limits.forgot.attempts', 3));
-        $this->ensureNotRateLimited($ipKey, config('jasmine.auth.rate_limits.forgot.ip_attempts', 10));
+        $this->ensureNotRateLimited($rlKey, (int)config('jasmine.auth.rate_limits.forgot.attempts', 3));
+        $this->ensureNotRateLimited($ipKey, (int)config('jasmine.auth.rate_limits.forgot.ip_attempts', 10));
 
         ResetPassword::createUrlUsing(function (Authenticatable $user, string $token) {
             return route('jasmine.password.reset', [
@@ -236,8 +236,8 @@ class AuthController extends Controller
 
         // Hit the limiter regardless of outcome — INVALID_USER must look identical to a real
         // send to prevent enumeration, and RESET_THROTTLED should still count against our cap.
-        RateLimiter::hit($rlKey, config('jasmine.auth.rate_limits.forgot.decay', 600));
-        RateLimiter::hit($ipKey, config('jasmine.auth.rate_limits.forgot.ip_decay', 600));
+        RateLimiter::hit($rlKey, (int)config('jasmine.auth.rate_limits.forgot.decay', 600));
+        RateLimiter::hit($ipKey, (int)config('jasmine.auth.rate_limits.forgot.ip_decay', 600));
 
         if ($res === Password::RESET_THROTTLED) {
             throw ValidationException::withMessages(['email' => [trans($res)]])
@@ -299,7 +299,7 @@ class AuthController extends Controller
         abort_unless((bool)$user->otp_secret, 404);
 
         $rlKey = "jasmine:otp|{$user->getAuthIdentifier()}";
-        $this->ensureNotRateLimited($rlKey, config('jasmine.auth.rate_limits.otp.attempts', 5), 'code');
+        $this->ensureNotRateLimited($rlKey, (int)config('jasmine.auth.rate_limits.otp.attempts', 5), 'code');
 
         $request->validate(['code' => [
             'required', 'digits:6',
@@ -309,7 +309,7 @@ class AuthController extends Controller
                 );
 
                 if ($timestamp === false) {
-                    RateLimiter::hit($rlKey, config('jasmine.auth.rate_limits.otp.decay', 900));
+                    RateLimiter::hit($rlKey, (int)config('jasmine.auth.rate_limits.otp.decay', 900));
                     $fail("The $attribute is invalid.");
 
                     return;
@@ -345,7 +345,7 @@ class AuthController extends Controller
         $user = static::guard()->user();
 
         $rlKey = "jasmine:webauthn|{$user->getAuthIdentifier()}";
-        $this->ensureNotRateLimited($rlKey, config('jasmine.auth.rate_limits.webauthn.attempts', 5), 'credential');
+        $this->ensureNotRateLimited($rlKey, (int)config('jasmine.auth.rate_limits.webauthn.attempts', 5), 'credential');
 
         $data = $request->validate(['credential' => ['required', 'array']]);
 
@@ -361,7 +361,7 @@ class AuthController extends Controller
         try {
             $webauthn->verifyAssertion($user, json_encode($data['credential']), $options);
         } catch (\Throwable) {
-            RateLimiter::hit($rlKey, config('jasmine.auth.rate_limits.webauthn.decay', 900));
+            RateLimiter::hit($rlKey, (int)config('jasmine.auth.rate_limits.webauthn.decay', 900));
             throw ValidationException::withMessages(['credential' => 'Could not verify the security key.']);
         }
